requirements.txt
absl-py==1.4.0
appdirs==1.4.4
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
decorator==4.4.2
docker-pycreds==0.4.0
farama-notifications==0.0.4
filelock==3.12.0
fonttools==4.38.0
gitdb==4.0.10
gitpython==3.1.31
google-auth-oauthlib==0.4.6
google-auth==2.18.0
grpcio==1.54.0
imageio-ffmpeg==0.3.0
imageio==2.28.1
importlib-metadata==5.2.0
jax-jumpy==1.0.0
kiwisolver==1.4.4
markdown==3.3.7
markupsafe==2.1.2
matplotlib==3.5.3
moviepy==1.0.3
numpy==1.21.6
oauthlib==3.2.2
packaging==23.1
pandas==1.3.5
pathtools==0.1.2
pillow==9.5.0
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.5
pyasn1-modules==0.3.0
pyasn1==0.5.0
pygame==2.1.0
pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
requests-oauthlib==1.3.1
requests==2.30.0
rich==11.2.0
rsa==4.7.2
sentry-sdk==1.22.2
setproctitle==1.3.2
setuptools==67.7.2
six==1.16.0
smmap==5.0.0
stable-baselines3==1.2.0
tenacity==8.2.3
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
torch==1.12.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
wandb==0.13.11
werkzeug==2.2.3
wheel==0.40.0
zipp==3.15.0
gymnasium==0.29.1
Cython==3.0.5
graphql-core==3.2.3
docker==7.0.0
seaborn~=0.13.1